separated the sidebar into two parts container/component. this should allow for utils to be used
