+++ app/components/Sidebar.js
@@ -4,10 +4,8 @@ import { connect } from 'react-redux';
 import styles from './Sidebar.css';
 import Project_List from './Project_List';
 import { fetchUserProjects } from '../reducers/projects_list';
-import * as Actions from '../../utilities/actions';
 
-export class Sidebar extends Component {
-
+export default class Sidebar extends Component {
   constructor(props) {
     super(props);
     this.onClickArchiveUpdate = this.onClickArchiveUpdate.bind(this);
@@ -59,25 +57,3 @@ export class Sidebar extends Component {
     );
   }
 }
-
-
-/* ---------------- CONTAINER --------------------*/
-function mapStateToProps(state) {
-  return {
-    loginUser: state.login,
-    projects: state.projects
-  }
-}
-
-function mapDispatchToProps(dispatch) {
-  return {
-    onLoadProjects: function (user) {
-      dispatch(fetchUserProjects(user));
-    }
-  }
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Sidebar);
